move stripe redirect and setPaymentByIntent into a dedicated hook
